Uses a running sum of the total number of cases
Uses a running sum of the total number of cases, running change, and assigns importance to each day based on the number of cases that come that day. 
Importance is assigned by using the number of cases on a particular day compared to the total number of cases.

Change in number of cases on day i  * importance of day 1+i gives the weighted value of the change in the ith day that will be averaged.